+++ src/engines/http.ts
@@ -1,8 +1,4 @@
-import {
-	ConnectionUnavailable,
-	HttpConnectionError,
-	MissingNamespaceDatabase,
-} from "../errors";
+import { ConnectionUnavailable, MissingNamespaceDatabase } from "../errors";
 import type { ExportOptions, RpcRequest, RpcResponse } from "../types";
 import { getIncrementalID } from "../util/get-incremental-id";
 import { retrieveRemoteVersion } from "../util/version-check";
@@ -179,49 +175,5 @@ export class HttpEngine extends AbstractEngine {
 
 		const dec = new TextDecoder("utf-8");
 		return dec.decode(buffer);
-	}
-
-	private async req_post(
-		body: unknown,
-		url?: URL,
-		headers_?: Record<string, string>,
-	): Promise<ArrayBuffer> {
-		const headers: Record<string, string> = {
-			"Content-Type": "application/cbor",
-			Accept: "application/cbor",
-			...headers_,
-		};
-
-		if (this.connection.namespace) {
-			headers["Surreal-NS"] = this.connection.namespace;
-		}
-
-		if (this.connection.database) {
-			headers["Surreal-DB"] = this.connection.database;
-		}
-
-		if (this.connection.token) {
-			headers.Authorization = `Bearer ${this.connection.token}`;
-		}
-
-		const raw = await fetch(`${url ?? this.connection.url}`, {
-			method: "POST",
-			headers,
-			body: this.encodeCbor(body),
-		});
-
-		const buffer = await raw.arrayBuffer();
-
-		if (raw.status === 200) {
-			return buffer;
-		}
-
-		const dec = new TextDecoder("utf-8");
-		throw new HttpConnectionError(
-			dec.decode(buffer),
-			raw.status,
-			raw.statusText,
-			buffer,
-		);
 	}
 }
